#19 copy-pasta serialized componentDiff, move doguDiff to own file
